requirements.txt
transformers>=4.30.0
datasets>=2.12.0
torch>=1.13.0
rouge-score>=0.1.2
streamlit>=1.20.0
evaluate>=0.4.0
matplotlib>=3.5.0
tqdm>=4.64.0
python-dotenv